requirements.txt
customtkinter
google-generativeai
packaging
requests
pillow
